Moved attackopts to models
