+++ internal/models/attack_params.go
@@ -61,14 +61,6 @@ const (
 	AttackResponseStatusFailed AttackStatus = "failed"
 )
 
-// AttackResponse with attacks UUID and AttackStatus
-type AttackResponse struct {
-	// id is a attack UUID generated for each attack submitted
-	ID string `json:"id,omitempty"`
-	// Status captures the attack status in the scheduler pipeline
-	Status AttackStatus `json:"status,omitempty"`
-}
-
 // AttackCancel
 type AttackCancel struct {
 	Cancel bool `json:"cancel"`
