+++ internal/models/attack_response.go
@@ -1,0 +1,9 @@
+package models
+
+// AttackResponse with attacks UUID and AttackStatus
+type AttackResponse struct {
+	// id is a attack UUID generated for each attack submitted
+	ID string `json:"id,omitempty"`
+	// Status captures the attack status in the scheduler pipeline
+	Status AttackStatus `json:"status,omitempty"`
+}
